# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMA API",
        "description": "Sistem Informasi Manajemen Amanah — REST API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base path"
        }
    ],
    "paths": {
...
        "/bank-reconciliations": {
            "get": {
                "tags": [
                    "BankReconciliation"
                ],
                "summary": "Daftar rekonsiliasi bank",
                "operationId": "eafe465a2d30228afd7024bbf2b0e0c8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "BankReconciliation"
                ],
                "summary": "Buat rekonsiliasi bank",
                "operationId": "d335a81446e017abe027445f582205e6",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bank-reconciliations/{bankReconciliation}": {
            "get": {
                "tags": [
                    "BankReconciliation"
                ],
                "summary": "Detail rekonsiliasi bank",
                "operationId": "ae350e130df6e782a7d1510984f656ae",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bank-reconciliations/{bankReconciliation}/lines": {
            "post": {
                "tags": [
                    "BankReconciliation"
                ],
                "summary": "Tambah baris rekonsiliasi",
                "operationId": "37093dbc326bab0a64e20515e680baa7",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/bank-reconciliations/{bankReconciliation}/complete": {
            "post": {
                "tags": [
                    "BankReconciliation"
                ],
                "summary": "Selesaikan rekonsiliasi bank",
                "operationId": "15600444c3674d0ec9d43d190cebe7d9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/funds": {
            "get": {
                "tags": [
                    "Fund"
                ],
                "summary": "Daftar Dana Amanah",
                "operationId": "4e73112f97cfd8c2725db81844efc260",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fund"
                ],
                "summary": "Buat Dana Amanah",
                "operationId": "b436b6ad3a866e8d102f4fa77e7f1f5c",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/funds/{fund}": {
            "get": {
                "tags": [
                    "Fund"
                ],
                "summary": "Detail Dana Amanah",
                "operationId": "dd24c61693d0dc922bb7d59813e7e70e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fund"
                ],
                "summary": "Ubah Dana Amanah",
                "operationId": "f2910652142d058dd4d5995de451aefd",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fund"
                ],
                "summary": "Nonaktifkan Dana Amanah",
                "operationId": "7c4d7e0ceba03b01426f9a7efb2c5db3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/programs": {
            "get": {
                "tags": [
                    "Program"
                ],
                "summary": "Daftar program",
                "operationId": "6f677f7a6f2a1759d1134cd8a19c0115",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Program"
                ],
                "summary": "Buat program",
                "operationId": "d89a30ece5239884ced56411f95ec746",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/programs/{program}": {
            "get": {
                "tags": [
                    "Program"
                ],
                "summary": "Detail program",
                "operationId": "e65e8baa7293e99b9fcf9df101ac4894",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Program"
                ],
                "summary": "Ubah program",
                "operationId": "a8117558534caf0fbfca4595b902eaa3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Program"
                ],
                "summary": "Nonaktifkan program",
                "operationId": "83a8615270c4737e79711c4d7f680bc2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "nullable": true
                    },
                    "meta": {
                        "nullable": true
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token dari POST /api/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}